pyproject.toml
[project]
name = "newchess"
version = "0.0.3"
authors = [{ name = "yourname", email = "your.name@TNO-or-deltares.nl" }]
description = "New CHESS"
classifiers = [
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Information Technology",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Mathematics",
    "Topic :: Scientific/Engineering :: Physics",
]

dependencies = [
    "pandapipes==0.8.4",
    "pyesdl==23.3.0",
    "python-dotenv==1.0.0",
    "peewee==3.16.2",
    "dataclass-wizard==0.22.2",
    "coloredlogs==15.0.1",
    "click==8.1.6",
]

[project.optional-dependencies]
dev = [
    "pip-tools~=6.13.0",
    "black~=22.1.0",
    "flake8==6.0.0",
    "Flake8-pyproject==1.2.3",
    "pytest ~=7.3.1",
    "pytest-cov ~=4.0.0",
    "bump2version==1.0.1",
]

doc = [
    "sphinx == 7.0.1",
    "furo == 2023.5.20",
    #    "sphinx-rtd-theme ~=1.2.2"
]

[project.urls]
homepage = "https://www.TNO-or-deltares.nl"
documentation = "https://readthedocs.org"
repository = "https://github.com"
changelog = "somethingsomethine"

[project.scripts]
newchess = "newchess.infrastructure.app:cmdline_app"

[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools ~= 67.7.2", "wheel ~= 0.40.0"]

[tools.setuptools]
packages = ["src/newchess"]

[tool.pytest.ini_options]
addopts = "--cov=your_package --cov-report html --cov-report term-missing --cov-fail-under 95"

[tool.coverage.run]
source = ["src"]

[tool.flake8]
exclude = ['venv/*', 'doc/*']
ignore = ['Q000', 'D401']
per-file-ignores = ['__init__.py:F401', 'test_main.py:D100,D101,D102,D103']
max-line-length = 100
count = true

[tool.black]
line-length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
exclude = [
    'venv/*',
    'doc/*',
    'build',
    'component',
    'testscript',
    'testmodel',
    'tryouts.py',
]

# mypy per-module options:
[[tool.mypy.overrides]]
module = "unit_test.*"
check_untyped_defs = true
ignore_missing_imports = true
